add tests for NonZeroI32
