+++ tooling/minimize/src/main.rs
@@ -28,6 +28,7 @@ mod rs {
     pub use rustc_span::source_map::Spanned;
     pub use rustc_span::{Span, DUMMY_SP};
     pub use rustc_target::abi::{self, call::*, Align, FieldIdx, Layout, Size};
+    pub use rustc_target::abi::{FieldsShape, TagEncoding, VariantIdx, Variants};
 }
 // Traits
 pub use rustc_middle::ty::layout::IntegerExt as _;
